Minor fix to enemy rectangle.

+++ Plane.cpp
@@ -16,7 +16,7 @@ const Rect Plane::getRect() const {
   #ifndef PLANES_HAVE_BORDERS
   return (Rect){ _x.getInteger(), _y.getInteger(), pgm_read_byte(pgm_read_word_near(_bitmaps)), pgm_read_byte(pgm_read_word_near(_bitmaps) + 1) };
   #else
-  return (Rect){ _x.getInteger() + 1, _y.getInteger() + 1, pgm_read_byte(pgm_read_word_near(_bitmaps)) - 2, pgm_read_byte(pgm_read_word_near(_bitmaps) + 1) - 2 };
+  return (Rect){ _x.getInteger() + static_cast<uint8_t>(1), _y.getInteger() + static_cast<uint8_t>(1), static_cast<uint8_t>(pgm_read_byte(pgm_read_word_near(_bitmaps))) - static_cast<uint8_t>(2), static_cast<uint8_t>(pgm_read_byte(pgm_read_word_near(_bitmaps) + 1)) - static_cast<uint8_t>(2) };
   #endif
 
 }
